T2990 partly completed.
Summary: Task is not finished yet. Its functionality is partly completed.

Test Plan:
This commit is for information only. The current solution lets the server work without a crash,
but the first transaction does not trigger revision management on a newly created model anyway (all changes are
applied though). The interesting part is that all following transactions DO trigger revision management
and everything works as supposed (at least it seems like that). It needs additional investigation why the
revision is not triggerd for the first time. Looks the system applies revision 1 when it is already one and
for the second time it applies 2 and it seems as if everything is okay. It seems like the revision management
system is behind for 1.

corresponds to hash: 128fd0a8ff0a194eb6ec4edc4904e37edf4d81b9